+++ src/views/LostItems/LostItemsListContainer/components/LostItemsList/components/RenderActions/components/ItemDetailsLink/ItemDetailsLink.js
@@ -9,7 +9,7 @@ import {
   ACTUAL_COST_RECORD_FIELD_NAME,
   ACTUAL_COST_RECORD_FIELD_PATH,
   DEFAULT_VALUE,
-} from '../../../../../../constants';
+} from '../../../../../../../constants';
 
 const ItemDetailsLink = ({ actualCostRecord }) => {
   const itemId = get(actualCostRecord, ACTUAL_COST_RECORD_FIELD_PATH[ACTUAL_COST_RECORD_FIELD_NAME.ITEM_ID], DEFAULT_VALUE);