+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix problem with remaining amount (not shown correct value after filling the payment amount). Refs UIU-2812.
 * Add "Status" column to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2772.
 * Add "Lost items requiring actual cost" to Actions dropdown in User details record. Refs UIU-2810.
+* Add "Fee/fine details" to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2773.
 
 ## [9.0.0](https://github.com/folio-org/ui-users/tree/v9.0.0) (2023-02-20)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v9.0.0)